+++ module/actor/maneuver-button.js
@@ -58,6 +58,8 @@ export default class ManeuverHUDButton {
    */
   static async prepTokenHUD(hud, html, token) {
     // const actor = this.getTokenActor(token)
+    if (!hud.object?.combatant) return
+
     const actor = hud.object?.actor
     const button = await this.createButton(token.effects)
 
@@ -80,8 +82,8 @@ export default class ManeuverHUDButton {
   // DONE add a migration to set the maneuver token effect for all tokens
   // DONE figure out how to remove maneuver from other status effects OR make clicking it add the "do nothing" maneuver
   // DONE implement the various options: \
-  //      - full detail: exact maneuver and subtype
-  //      - general detail: maneuver name w/o subtype
+  //      √ full detail: exact maneuver and option
+  //      √ general detail: maneuver name w/o option
   // TODO implement visibility: \
   //      √ everyone
   //      √ GM and Owner only
@@ -90,4 +92,5 @@ export default class ManeuverHUDButton {
   // TODO Ultimately turn this into an Active Effect - and actually adjust Move and/or other conditions
   // TODO Add status hint text to modifier bucket
   // DONE Make sure Tokens are initialized with a Maneuver
+  // DONE Add maneuver when token enters combat; remove when not in combat
 }