+++ module/actor/maneuver.js
@@ -26,6 +26,7 @@ const ManeuverData = {
     defense: DEFENSE_ANY,
     icon: 'systems/gurps/icons/maneuvers/man-nothing.png',
     label: 'GURPS.maneuverDoNothing',
+    fullturn: false,
   },
   move: {
     id: 'move',
@@ -33,6 +34,7 @@ const ManeuverData = {
     defense: DEFENSE_ANY,
     icon: 'systems/gurps/icons/maneuvers/man-move.png',
     label: 'GURPS.maneuverMove',
+    fullturn: false,
   },
   aim: {
     id: 'aim',
@@ -40,6 +42,7 @@ const ManeuverData = {
     defense: DEFENSE_ANY,
     icon: 'systems/gurps/icons/maneuvers/man-aim.png',
     label: 'GURPS.maneuverAim',
+    fullturn: true,
   },
   change_posture: {
     id: 'change_posture',
@@ -47,6 +50,7 @@ const ManeuverData = {
     defense: DEFENSE_ANY,
     icon: 'systems/gurps/icons/maneuvers/man-change-posture.png',
     label: 'GURPS.maneuverChangePosture',
+    fullturn: false,
   },
   evaluate: {
     id: 'evaluate',
@@ -54,6 +58,7 @@ const ManeuverData = {
     defense: DEFENSE_ANY,
     icon: 'systems/gurps/icons/maneuvers/man-evaluate.png',
     label: 'GURPS.maneuverEvaluate',
+    fullturn: false,
   },
   attack: {
     id: 'attack',
@@ -61,6 +66,7 @@ const ManeuverData = {
     defense: DEFENSE_ANY,
     icon: 'systems/gurps/icons/maneuvers/man-attack.png',
     label: 'GURPS.maneuverAttack',
+    fullturn: false,
   },
   feint: {
     id: 'feint',
@@ -69,6 +75,7 @@ const ManeuverData = {
     icon: 'systems/gurps/icons/maneuvers/man-feint.png',
     alt: 'systems/gurps/icons/maneuvers/man-attack.png',
     label: 'GURPS.maneuverFeint',
+    fullturn: false,
   },
   allout_attack: {
     id: 'allout_attack',
@@ -76,6 +83,7 @@ const ManeuverData = {
     defense: DEFENSE_NONE,
     icon: 'systems/gurps/icons/maneuvers/man-allout-attack.png',
     label: 'GURPS.maneuverAllOutAttack',
+    fullturn: false,
   },
   aoa_determined: {
     id: 'aoa_determined',
@@ -84,6 +92,7 @@ const ManeuverData = {
     icon: 'systems/gurps/icons/maneuvers/man-aoa-determined.png',
     alt: 'systems/gurps/icons/maneuvers/man-allout-attack.png',
     label: 'GURPS.maneuverAllOutAttackDetermined',
+    fullturn: false,
   },
   aoa_double: {
     id: 'aoa_double',
@@ -92,6 +101,7 @@ const ManeuverData = {
     icon: 'systems/gurps/icons/maneuvers/man-aoa-double.png',
     alt: 'systems/gurps/icons/maneuvers/man-allout-attack.png',
     label: 'GURPS.maneuverAllOutAttackDouble',
+    fullturn: false,
   },
   aoa_feint: {
     id: 'aoa_feint',
@@ -100,6 +110,7 @@ const ManeuverData = {
     icon: 'systems/gurps/icons/maneuvers/man-aoa-feint.png',
     alt: 'systems/gurps/icons/maneuvers/man-allout-attack.png',
     label: 'GURPS.maneuverAllOutAttackFeint',
+    fullturn: false,
   },
   aoa_strong: {
     id: 'aoa_strong',
@@ -108,6 +119,7 @@ const ManeuverData = {
     icon: 'systems/gurps/icons/maneuvers/man-aoa-strong.png',
     alt: 'systems/gurps/icons/maneuvers/man-allout-attack.png',
     label: 'GURPS.maneuverAllOutAttackStrong',
+    fullturn: false,
   },
   aoa_suppress: {
     id: 'aoa_suppress',
@@ -116,6 +128,7 @@ const ManeuverData = {
     icon: 'systems/gurps/icons/maneuvers/man-aoa-suppress.png',
     alt: 'systems/gurps/icons/maneuvers/man-allout-attack.png',
     label: 'GURPS.maneuverAllOutAttackSuppressFire',
+    fullturn: true,
   },
   move_and_attack: {
     id: 'move_and_attack',
@@ -123,6 +136,7 @@ const ManeuverData = {
     defense: DEFENSE_DODGEBLOCK,
     icon: 'systems/gurps/icons/maneuvers/man-move-attack.png',
     label: 'GURPS.maneuverMoveAttack',
+    fullturn: false,
   },
   allout_defense: {
     id: 'allout_defense',
@@ -130,6 +144,7 @@ const ManeuverData = {
     defense: DEFENSE_ANY,
     icon: 'systems/gurps/icons/maneuvers/man-defense.png',
     label: 'GURPS.maneuverAllOutDefense',
+    fullturn: false,
   },
   aod_dodge: {
     id: 'aod_dodge',
@@ -138,6 +153,7 @@ const ManeuverData = {
     icon: 'systems/gurps/icons/maneuvers/man-def-dodge.png',
     alt: 'systems/gurps/icons/maneuvers/man-defense.png',
     label: 'GURPS.maneuverAllOutDefenseDodge',
+    fullturn: false,
   },
   aod_parry: {
     id: 'aod_parry',
@@ -146,6 +162,7 @@ const ManeuverData = {
     icon: 'systems/gurps/icons/maneuvers/man-def-parry.png',
     alt: 'systems/gurps/icons/maneuvers/man-defense.png',
     label: 'GURPS.maneuverAllOutDefenseParry',
+    fullturn: false,
   },
   aod_block: {
     id: 'aod_block',
@@ -154,6 +171,7 @@ const ManeuverData = {
     icon: 'systems/gurps/icons/maneuvers/man-def-block.png',
     alt: 'systems/gurps/icons/maneuvers/man-defense.png',
     label: 'GURPS.maneuverAllOutDefenseBlock',
+    fullturn: false,
   },
   aod_double: {
     id: 'aod_double',
@@ -162,6 +180,7 @@ const ManeuverData = {
     icon: 'systems/gurps/icons/maneuvers/man-def-double.png',
     alt: 'systems/gurps/icons/maneuvers/man-defense.png',
     label: 'GURPS.maneuverAllOutDefenseDouble',
+    fullturn: false,
   },
   ready: {
     id: 'ready',
@@ -169,6 +188,7 @@ const ManeuverData = {
     defense: DEFENSE_ANY,
     icon: 'systems/gurps/icons/maneuvers/man-ready.png',
     label: 'GURPS.maneuverReady',
+    fullturn: false,
   },
   concentrate: {
     id: 'concentrate',
@@ -176,6 +196,7 @@ const ManeuverData = {
     defense: DEFENSE_ANY,
     icon: 'systems/gurps/icons/maneuvers/man-concentrate.png',
     label: 'GURPS.maneuverConcentrate',
+    fullturn: true,
   },
   wait: {
     id: 'wait',
@@ -183,6 +204,7 @@ const ManeuverData = {
     defense: DEFENSE_ANY,
     icon: 'systems/gurps/icons/maneuvers/man-wait.png',
     label: 'GURPS.maneuverWait',
+    fullturn: false,
   },
 }
 
@@ -232,6 +254,7 @@ export class Maneuvers {
   }
 }
 
+// Our override of the TokenHUD; it removes the maneuver tokens from the list of status effects
 export class GURPSTokenHUD extends TokenHUD {
   getData(options) {
     let data = super.getData(options)
@@ -248,6 +271,7 @@ export class GURPSTokenHUD extends TokenHUD {
 }
 
 Hooks.once('init', () => {
+  // Patch the method Token#drawEffects so we can monkey with the token effects before drawing
   const original_Token_drawEffects = Token.prototype.drawEffects
   Token.prototype.drawEffects = async function (...args) {
     const tokenEffects = this.data.effects
@@ -281,12 +305,35 @@ Hooks.once('init', () => {
       }
     }
 
+    // call the original method
     const result = original_Token_drawEffects.apply(this, args)
 
+    // restore the original token effects
     this.data.effects = tokenEffects
 
     return result
   }
 })
 
-// export function Token_drawEffect(src, i, bg, w, tint)
+// on create combatant, set the maneuver
+Hooks.on('createCombatant', (combat, options, id) => {
+  console.log(id)
+  let token = combat.token
+  let actor = combat.actor
+  actor.updateManeuver('do_nothing', token.id)
+})
+
+// on delete combatant, remove the maneuver
+Hooks.on('deleteCombatant', (combat, options, id) => {
+  console.log(id)
+  let token = combat.token
+  let actor = combat.actor
+  actor.removeManeuver(token.id)
+})
+
+Hooks.on('deleteCombat', (combat, options, id) => {
+  let combatants = combat.data.combatants.contents
+  combatants.forEach(it => it.actor.removeManeuver(it.token.id))
+})
+
+// TODO consider subtracting 1 FP from every combatant that leaves combat